add eslint
