+++ src/games/Board/Board.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import { Box } from '@material-ui/core';
 import Tile from '../Tile/Tile';
 import { useDispatch } from 'react-redux';
-import { clearBoard, getPieces } from '../../store/activeGame/activeGame';
+import { clearBoard } from '../../store/activeGame/activeGame';
+import { getPieces } from '../../store/activeGame/activeGame.thunk';
 import { usePieceSize } from '../../utils/hooks';
 
 interface Props {
@@ -10,7 +11,7 @@ interface Props {
 }
 
 const Board: React.FC<Props> = ({ gameId }) => {
-  const array = new Array(8).fill(0);
+  const array = Array.from(Array(8).keys());
   const dispatch = useDispatch();
   const pieceSize = usePieceSize();
 
@@ -27,9 +28,7 @@ const Board: React.FC<Props> = ({ gameId }) => {
       gridTemplateColumns={`repeat(8, ${pieceSize})`}
       gridTemplateRows={`repeat(8, ${pieceSize})`}
     >
-      {array.map((_, row) =>
-        array.map((__, col) => <Tile key={`tile-${row}-${col}`} row={row} col={col} />)
-      )}
+      {array.map(row => array.map(col => <Tile key={`tile-${row}-${col}`} row={row} col={col} />))}
     </Box>
   );
 };
