+++ src/information/CreateGameButton/Attributions.tsx
@@ -20,7 +20,7 @@ const Attributions: React.FC = () => {
 
   return (
     <Typography className={classes.footer}>
-      Icons by en:User:Cburnett - Own work  This W3C-unspecified vector image was created with
+      Icons by en:User:Cburnett - Own work This W3C-unspecified vector image was created with
       Inkscape., CC BY-SA 3.0, https://commons.wikimedia.org/w/index.php?curid=1499810
     </Typography>
   );
