+++ src/services/Api.ts
@@ -1,6 +1,6 @@
-import axios from 'axios'
-import config from '../config'
+import axios from 'axios';
+import config from '../config';
 
 export const chessApi = axios.create({
-    baseURL: config.serviceUrl
-})
+  baseURL: config.serviceUrl
+});
